src/IRDose/requirements.txt
certifi==2018.4.16
cloudinary==1.12.0
dj-database-url==0.5.0
Django==2.0.4
django-crispy-forms==1.7.2
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.7.1
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
pandas==0.22.0
pbr==4.2.0
Pillow==5.1.0
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
scipy==1.0.1
six==1.11.0
urllib3==1.23
vtk==8.1.1
